Moved the location of inputs file from Helpers to Data and Insight folder
